fix setter parameters

+++ core/src/ca/bcit/comp2522/termproject/Entity.java
@@ -40,7 +40,7 @@ abstract public class Entity {
         this.attack = attack;
     }
 
-    public void setDefense(int defense) {
+    public void setDefense(double defense) {
         this.defense = defense;
     }
